requirements.txt
playwright>=1.40.0
playwright-stealth>=1.0.6
pywebview>=5.0
pydantic>=2.0.0
mcp
# macOS WebView support
pyobjc-framework-WebKit; sys_platform == 'darwin'
pyobjc-core; sys_platform == 'darwin'
